requirements.txt
bentoml
joblib
numpy
pandas
pathlib
pip
pydantic
pyjwt
pytest
requests
scikit-learn
xgboost